requirements.txt
streamlit>=1.20.0
streamlit-cropper>=0.2.2
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0
Pillow>=9.4.0
